+++ tbot/tc/selftest/tc/__init__.py
@@ -7,6 +7,7 @@ from .git import *  # noqa: F403
 from .shell import *  # noqa: F403
 from .build import *  # noqa: F403
 from .uboot import *  # noqa: F403
+from .kconfig import *  # noqa: F403
 
 
 @tbot.testcase
@@ -23,5 +24,6 @@ def selftest_tc(lab: typing.Optional[linux.LabHost] = None,) -> None:
             selftest_tc_uboot_checkout,  # noqa: F405
             selftest_tc_uboot_build,  # noqa: F405
             selftest_tc_uboot_patched_bisect,  # noqa: F405
+            selftest_tc_kconfig,  # noqa: F405
             lab=lh,
         )